python-bot/requirements.txt
# requirements.txt
pyTelegramBotAPI==4.22.1
websocket-client==1.4.1
python-telegram-bot==20.0
python-dotenv==0.19.0 
